requirements.txt
setuptools
python-decouple
psycopg2-binary
fastapi
uvicorn
requests
selenium
